:bug: 전역 상태값에 따른 모달 분기별 리팩토링
